Got a very RAW authorization scheme, dont like it but it work for now

+++ Monitoring/Controllers/SensorsController.cs
@@ -1,15 +1,19 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using AutoMapper;
+using Dapper;
 using DTO;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Authorization.Infrastructure;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Models;
 using Models.Interfaces;
 using Models.Repository;
+using Services.Interfaces;
 
 namespace Monitoring.Controllers
 {
@@ -20,11 +24,13 @@ namespace Monitoring.Controllers
     {
         private readonly ISensorsRepository _sensorsRepository;
         private readonly IMapper _mapper;
+        private readonly ISensorAuthorizationService _authorizationService;
 
-        public SensorsController(ISensorsRepository iSensorsRepository, IMapper mapper)
+        public SensorsController(ISensorsRepository iSensorsRepository, IMapper mapper, ISensorAuthorizationService authorizationService)
         {
             _sensorsRepository = iSensorsRepository;
             _mapper = mapper;
+            _authorizationService = authorizationService;
         }
 
         //GET api/sensors
@@ -45,13 +51,19 @@ namespace Monitoring.Controllers
         [HttpGet("{id}")]
         public async Task<IActionResult> GetSensor(int id)
         {
-            Sensor sensor = await _sensorsRepository.GetById(id);
+            var userId = HttpContext.User.Claims.SingleOrDefault(x => x.Type == ClaimTypes.NameIdentifier)?.Value;
+            if (userId == null)
+                return BadRequest();
 
+            bool found = await _authorizationService.IsAuthorized(int.Parse(userId), id, Permissions.View);
+            if (!found)
+                return Unauthorized();
+
+            Sensor sensor = await _sensorsRepository.GetById(id);
             if (sensor == null)
                 return NotFound();
 
             var sensorDto = _mapper.Map<Sensor, SensorDto>(sensor);
-
             return Ok(sensorDto);
         }
 
@@ -59,6 +71,14 @@ namespace Monitoring.Controllers
         [HttpPost]
         public async Task<IActionResult> Post([FromBody] SensorDto sensorDto)
         {
+            var userId = HttpContext.User.Claims.SingleOrDefault(x => x.Type == ClaimTypes.NameIdentifier)?.Value;
+            if (userId == null)
+                return BadRequest();
+
+            bool found = await _authorizationService.IsAuthorized(int.Parse(userId), sensorDto.Id, Permissions.View);
+            if (!found)
+                return Unauthorized();
+
             Sensor sensor = _mapper.Map<SensorDto, Sensor>(sensorDto);
             int result = await _sensorsRepository.Create(sensor);
 
@@ -72,9 +92,17 @@ namespace Monitoring.Controllers
         [HttpPut("{id}")]
         public async Task<IActionResult> Put(int id, [FromBody] SensorDto sensorDto)
         {
-            Sensor sensor = _mapper.Map<SensorDto, Sensor>(sensorDto);
-            bool success = await _sensorsRepository.Update(sensor);
+            var userId = HttpContext.User.Claims.SingleOrDefault(x => x.Type == ClaimTypes.NameIdentifier)?.Value;
 
+            if (userId == null)
+                return BadRequest();
+            bool found = await _authorizationService.IsAuthorized(int.Parse(userId), sensorDto.Id, Permissions.View);
+
+            if (!found)
+                return Unauthorized();
+            Sensor sensor = _mapper.Map<SensorDto, Sensor>(sensorDto);
+
+            bool success = await _sensorsRepository.Update(sensor);
             if (!success)
                 return UnprocessableEntity();
 
@@ -85,6 +113,14 @@ namespace Monitoring.Controllers
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(int id)
         {
+            var userId = HttpContext.User.Claims.SingleOrDefault(x => x.Type == ClaimTypes.NameIdentifier)?.Value;
+
+            if (userId == null)
+                return BadRequest();
+            bool found = await _authorizationService.IsAuthorized(int.Parse(userId), id, Permissions.View);
+
+            if (!found)
+                return Unauthorized();
             bool success = await _sensorsRepository.Delete(id);
 
             if (!success)
